Fixed PR comments.
