+++ animals/src/main/java/animals/enclosures/EnclosureClone.java
@@ -1,0 +1,14 @@
+package animals.enclosures;
+
+import animals.model.mammals.artiodactyls.Cow;
+
+public class EnclosureClone {
+
+    public Enclosure<Cow> cloneEnclosureOfCows(Enclosure<Cow> enclosureToClone) throws CloneNotSupportedException {
+        Enclosure<Cow> clonedEnclosure = new Enclosure<>();
+        for (Cow cow : enclosureToClone.getAnimals()) {
+            clonedEnclosure.addAnimal(cow.clone());
+        }
+        return clonedEnclosure;
+    }
+}
