+++ animals/src/main/java/animals/exception/IncorrectGroundMobilityRangeException.java
@@ -1,0 +1,8 @@
+package animals.exception;
+
+public class IncorrectGroundMobilityRangeException extends RuntimeException {
+
+    public IncorrectGroundMobilityRangeException(String message) {
+        super(message);
+    }
+}
