+++ animals/src/main/java/animals/model/Animal.java
@@ -1,0 +1,45 @@
+package animals.model;
+
+import java.util.Objects;
+
+public abstract class Animal {
+    private int age;
+    private int weight;
+
+    public Animal(int age, int weight) {
+        this.age = age;
+        this.weight = weight;
+    }
+
+    public Animal() {
+    }
+
+    public int getAge() {
+        return age;
+    }
+
+    public void setAge(int age) {
+        this.age = age;
+    }
+
+    public int getWeight() {
+        return weight;
+    }
+
+    public void setWeight(int weight) {
+        this.weight = weight;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        Animal animal = (Animal) o;
+        return age == animal.age && weight == animal.weight;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(age, weight);
+    }
+}
