+++ animals/src/main/java/animals/model/birds/Bird.java
@@ -1,0 +1,13 @@
+package animals.model.birds;
+
+import animals.model.Animal;
+
+public abstract class Bird extends Animal {
+
+    public Bird(int age, int weight) {
+        super();
+    }
+
+    public Bird() {
+    }
+}
