+++ animals/src/main/java/animals/model/mammals/Mammal.java
@@ -1,0 +1,13 @@
+package animals.model.mammals;
+
+import animals.model.Animal;
+
+public abstract class Mammal extends Animal {
+
+    public Mammal(int age, int weight) {
+        super(age, weight);
+    }
+
+    public Mammal() {
+    }
+}
