+++ animals/src/main/java/animals/model/mammals/artiodactyls/Horse.java
@@ -1,0 +1,12 @@
+package animals.model.mammals.artiodactyls;
+
+public class Horse extends Artiodactyl {
+
+    public Horse(int age, int weight, boolean hasHorns) {
+        super(age, weight, hasHorns);
+    }
+
+    public Horse() {
+        super();
+    }
+}
