+++ animals/src/main/java/animals/model/mammals/predators/aquatic/Cachalot.java
@@ -1,0 +1,12 @@
+package animals.model.mammals.predators.aquatic;
+
+public class Cachalot extends AquaticPredator {
+
+    public Cachalot(int age, int weight, WeaponType weaponType, int minWaterTemperature) {
+        super(age, weight, weaponType, minWaterTemperature);
+    }
+
+    public Cachalot() {
+        super();
+    }
+}
