+++ animals/src/main/java/animals/model/mammals/predators/aquatic/KillerWhale.java
@@ -1,0 +1,12 @@
+package animals.model.mammals.predators.aquatic;
+
+public class KillerWhale extends AquaticPredator {
+
+    public KillerWhale(int age, int weight, WeaponType weaponType, int minWaterTemperature) {
+        super(age, weight, weaponType, minWaterTemperature);
+    }
+
+    public KillerWhale() {
+        super();
+    }
+}
