+++ animals/src/main/java/animals/model/mammals/predators/ground/Coyote.java
@@ -1,0 +1,12 @@
+package animals.model.mammals.predators.ground;
+
+public class Coyote extends GroundPredator {
+
+    public Coyote(int age, int weight, WeaponType weaponType, int groundMobility) {
+        super(age, weight, weaponType, groundMobility);
+    }
+
+    public Coyote() {
+        super();
+    }
+}
