+++ animals/src/main/java/animals/model/mammals/predators/ground/Lion.java
@@ -1,0 +1,12 @@
+package animals.model.mammals.predators.ground;
+
+public class Lion extends GroundPredator {
+
+    public Lion(int age, int weight, WeaponType weaponType, int groundMobility) {
+        super(age, weight, weaponType, groundMobility);
+    }
+
+    public Lion() {
+        super();
+    }
+}
