+++ animals/src/main/java/animals/model/mammals/predators/mixed/WhiteBear.java
@@ -1,0 +1,13 @@
+package animals.model.mammals.predators.mixed;
+
+public class WhiteBear extends MixedPredator {
+
+    public WhiteBear(int age, int weight, WeaponType weaponType,
+                     PreferredHuntingEnvironment preferredHuntingEnvironment) {
+        super(age, weight, weaponType, preferredHuntingEnvironment);
+    }
+
+    public WhiteBear() {
+        super();
+    }
+}
