+++ animals/src/main/java/animals/model/reptiles/Reptile.java
@@ -1,0 +1,13 @@
+package animals.model.reptiles;
+
+import animals.model.Animal;
+
+public abstract class Reptile extends Animal {
+
+    public Reptile(int age, int weight) {
+        super(age, weight);
+    }
+
+    public Reptile() {
+    }
+}
