+++ animals/src/main/java/animals/model/spiders/Spider.java
@@ -1,0 +1,13 @@
+package animals.model.spiders;
+
+import animals.model.Animal;
+
+public abstract class Spider extends Animal {
+
+    public Spider(int age, int weight) {
+        super(age, weight);
+    }
+
+    public Spider() {
+    }
+}
