--- animals/src/main/java/database/exception/IncorrectGroundMobilityRangeException.java
@@ -1,8 +1,0 @@
-package database.exception;
-
-public class IncorrectGroundMobilityRangeException extends RuntimeException {
-
-    public IncorrectGroundMobilityRangeException(String message) {
-        super(message);
-    }
-}
