--- animals/src/main/java/practice/model/birds/Bird.java
@@ -1,13 +1,0 @@
-package practice.model.birds;
-
-import practice.model.Animal;
-
-public abstract class Bird extends Animal {
-
-    public Bird(int age, int weight) {
-        super();
-    }
-
-    public Bird() {
-    }
-}
