--- animals/src/main/java/practice/model/mammals/Mammal.java
@@ -1,13 +1,0 @@
-package practice.model.mammals;
-
-import practice.model.Animal;
-
-public abstract class Mammal extends Animal {
-
-    public Mammal(int age, int weight) {
-        super(age, weight);
-    }
-
-    public Mammal() {
-    }
-}
