--- animals/src/main/java/practice/model/mammals/artiodactyls/Cow.java
@@ -1,17 +1,0 @@
-package practice.model.mammals.artiodactyls;
-
-public class Cow extends Artiodactyl implements Cloneable {
-
-    public Cow(int age, int weight, boolean hasHorns) {
-        super(age, weight, hasHorns);
-    }
-
-    public Cow() {
-        super();
-    }
-
-    @Override
-    public Cow clone() throws CloneNotSupportedException {
-        return (Cow) super.clone();
-    }
-}
