--- animals/src/main/java/practice/model/mammals/artiodactyls/Horse.java
@@ -1,12 +1,0 @@
-package practice.model.mammals.artiodactyls;
-
-public class Horse extends Artiodactyl {
-
-    public Horse(int age, int weight, boolean hasHorns) {
-        super(age, weight, hasHorns);
-    }
-
-    public Horse() {
-        super();
-    }
-}
