--- animals/src/main/java/practice/model/mammals/predators/aquatic/Cachalot.java
@@ -1,12 +1,0 @@
-package practice.model.mammals.predators.aquatic;
-
-public class Cachalot extends AquaticPredator {
-
-    public Cachalot(int age, int weight, WeaponType weaponType, int minWaterTemperature) {
-        super(age, weight, weaponType, minWaterTemperature);
-    }
-
-    public Cachalot() {
-        super();
-    }
-}
