--- animals/src/main/java/practice/model/mammals/predators/aquatic/KillerWhale.java
@@ -1,12 +1,0 @@
-package practice.model.mammals.predators.aquatic;
-
-public class KillerWhale extends AquaticPredator {
-
-    public KillerWhale(int age, int weight, WeaponType weaponType, int minWaterTemperature) {
-        super(age, weight, weaponType, minWaterTemperature);
-    }
-
-    public KillerWhale() {
-        super();
-    }
-}
