--- animals/src/main/java/practice/model/mammals/predators/ground/Coyote.java
@@ -1,12 +1,0 @@
-package practice.model.mammals.predators.ground;
-
-public class Coyote extends GroundPredator {
-
-    public Coyote(int age, int weight, WeaponType weaponType, int groundMobility) {
-        super(age, weight, weaponType, groundMobility);
-    }
-
-    public Coyote() {
-        super();
-    }
-}
