--- animals/src/main/java/practice/model/mammals/predators/ground/Lion.java
@@ -1,12 +1,0 @@
-package practice.model.mammals.predators.ground;
-
-public class Lion extends GroundPredator {
-
-    public Lion(int age, int weight, WeaponType weaponType, int groundMobility) {
-        super(age, weight, weaponType, groundMobility);
-    }
-
-    public Lion() {
-        super();
-    }
-}
