--- animals/src/main/java/practice/model/mammals/predators/mixed/JungleCat.java
@@ -1,13 +1,0 @@
-package practice.model.mammals.predators.mixed;
-
-public class JungleCat extends MixedPredator {
-
-    public JungleCat(int age, int weight, WeaponType weaponType,
-                     PreferredHuntingEnvironment preferredHuntingEnvironment) {
-        super(age, weight, weaponType, preferredHuntingEnvironment);
-    }
-
-    public JungleCat() {
-        super();
-    }
-}
