--- animals/src/main/java/practice/model/reptiles/Reptile.java
@@ -1,13 +1,0 @@
-package practice.model.reptiles;
-
-import practice.model.Animal;
-
-public abstract class Reptile extends Animal {
-
-    public Reptile(int age, int weight) {
-        super(age, weight);
-    }
-
-    public Reptile() {
-    }
-}
