--- animals/src/main/java/practice/model/spiders/Spider.java
@@ -1,13 +1,0 @@
-package practice.model.spiders;
-
-import practice.model.Animal;
-
-public abstract class Spider extends Animal {
-
-    public Spider(int age, int weight) {
-        super(age, weight);
-    }
-
-    public Spider() {
-    }
-}
